usize -> u32 in spans
to cut the memory usage of parsing by roughly half + give us a bit of a speed bonus

+++ muscript-parsing/src/ast/expr.rs
@@ -498,7 +498,7 @@ impl Token {
     fn is_compound_assignment(&self, input: &str) -> bool {
         // This is a little bit cursed, but the level of cursedness here is nothing compared to
         // UnrealScript as a whole.
-        input[self.span.end..].starts_with('=')
+        input[self.span.end as usize..].starts_with('=')
     }
 }
